# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Road Trip Bingo Word Database API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/autofill:
    post:
      operationId: autofillWords
      tags: [ai]
      summary: Use AI to autofill missing tags for words
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutofillRequest"
      responses:
        "200":
          description: Autofill results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutofillResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Another AI request is already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/suggest:
    post:
      operationId: suggestWords
      tags: [ai]
      summary: Use AI to suggest new words to add to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestWordsRequest"
      responses:
        "200":
          description: Suggested words
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestWordsResponse"
        "429":
          description: Another AI request is already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Word:
      type: object
      properties:
        id:
          type: integer
        word:
          type: string
        regions:
          type: array
          items:
            type: string
          description: "e.g. All, NE, SE, N Cent, S Cent, NW, SW, HI, AK"
        surroundings:
          type: array
          items:
            type: string
          description: "e.g. All, Rural/Xurban, Suburban/Town, Urban/City, Highway"
        dayNight:
          type: array
          items:
            type: string
          description: "e.g. Day, Night"
        age:
          type: ["string", "null"]
          description: "e.g. Young, Kid, Tween"
        findability:
          type: ["string", "null"]
          description: "e.g. High, Medium, Low"
        seasons:
          type: array
          items:
            type: string
          description: "e.g. All, Spring, Summer, Fall, Winter"
        boards:
          type: array
          items:
            type: string
          description: "Board names (read-only, derived from boardIds)"
        boardIds:
          type: array
          items:
            type: integer
          description: "IDs of boards this word is associated with"
        notes:
          type: ["string", "null"]
        spanish:
          type: ["string", "null"]
          description: "Mexican Spanish translation"
        emoji:
          type: ["string", "null"]
          description: "System emoji representing the word"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - word
        - regions
        - surroundings
        - dayNight
        - age
        - findability
        - seasons
        - boards
        - boardIds
        - notes
        - spanish
        - emoji
        - createdAt
        - updatedAt
        - deletedAt
...
    AutofillRequest:
      type: object
      properties:
        wordIds:
          type: array
          items:
            type: integer
          maxItems: 50
          description: IDs of words to autofill. If empty, autofills all incomplete words (capped at 50 per batch).
        fields:
          type: array
          items:
            type: string
          maxItems: 7
          description: "Which fields to autofill: regions, surroundings, dayNight, age, findability, seasons, boards"
      required:
        - fields
    AutofillResponse:
      type: object
      properties:
        updated:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/Word"
      required:
        - updated
        - results
    SuggestWordsRequest:
      type: object
      properties:
        theme:
          type: ["string", "null"]
          maxLength: 200
        count:
          type: integer
          default: 10
          minimum: 1
          maximum: 20
      required: []
    SuggestWordsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              word:
                type: string
              rationale:
                type: string
            required:
              - word
              - rationale
      required:
        - suggestions
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Static API key for the public /v1 namespace. This key ships inside the
        iOS app binary — it is attribution/abuse control, not real access
        control. Admin routes use a separate password-gated bearer token.